layers: Rename Assistant layer files
