+++ layer_factory/assistant_layer/mem_objects.h
